finish clanmembers and levels

+++ src/main/java/com/runescape/info/model/Level.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Level {
     private Skill skill;
-    private Integer rank;
-    private Integer level;
-    private Integer experience;
+    private Long rank;
+    private Long level;
+    private Long experience;
 }
